+++ src/components/Player.jsx
@@ -5,7 +5,7 @@ import { useFrame } from "@react-three/fiber";
 import { useControls, folder } from "leva";
 import * as THREE from "three/webgpu";
 import {mrt, uniform, output} from "three/tsl";
-import buildNodeMaterialFromExisting from "./stores/useNodeMaterialConverter.jsx";
+import buildNodeMaterialFromExisting from "../stores/useNodeMaterialConverter.jsx";
 
 const UP_VECTOR = new THREE.Vector3(0, 1, 0);
